FP-51 Move ExportShotsThread and LoadShotsThread to own file
