+++ python/app/models/__init__.py
@@ -4,3 +4,5 @@ from .NukeProcess import NukeProcess
 from .Deliverables import Deliverables
 from .Settings import Settings
 from .UserSettings import UserSettings
+from .ExportShotsThread import ExportShotsThread
+from .LoadShotsThread import LoadShotsThread
